chore(db): Update UserDAO to use query_as! and use i64 IDs in most places

+++ ee/tabby-db/src/lib.rs
@@ -355,7 +355,7 @@ mod tests {
 pub mod testutils {
     use super::*;
 
-    pub async fn create_user(conn: &DbConn) -> i32 {
+    pub async fn create_user(conn: &DbConn) -> i64 {
         let email: &str = "test@example.com";
         let password: &str = "123456789";
         conn.create_user(email.to_string(), Some(password.to_string()), true)
